requirements.txt
aiofiles
mysql-connector-python
aioredis
alembic
amqp
anyio
async-timeout
attrs
autoflake
Babel
bcrypt
billiard
certifi
cffi
click
click-didyoumean
click-plugins
async-generator
click-repl
cryptography
dnspython
email-validator
exceptiongroup
fastapi
fastapi-cache2
fastapi-users
fastapi-users-db-sqlalchemy
greenlet
h11
httpcore
httptools
httpx
importlib-metadata
importlib-resources
iniconfig
iso8601
itsdangerous
kombu
makefun
Mako
MarkupSafe
outcome
passlib
pendulum
pluggy
prompt-toolkit
psycopg2-binary
pydantic
pyflakes
pycparser
packaging
idna
PyJWT
pypika-tortoise
pytest
pytest-asyncio
python-dateutil
python-dotenv
python-multipart
pytz
pytzdata
PyYAML
redis
rfc3986
six
sniffio
sortedcontainers
SQLAlchemy
sqlalchemy_database
starlette
tomli
tortoise-orm
typing_extensions
ujson
uvicorn
vine
watchfiles
wcwidth
WTForms
zipp
flower
websockets==11.0.3
gunicorn
sqlalchemy2-stubs
fastapi_profiler
sqladmin==0.11.0
numpy
aiohttp
token-throttler
psutil
